CLDR-16758 Too many abstained votes; fix bug in import votes
-Move equalsOrInheritsCurrentValue from XMLSource to CLDRFile

-The callers had non-resolving XMLSource for which getBaileyValue is always null
