+++ tools/cldr-code/src/main/java/org/unicode/cldr/util/CLDRFile.java
@@ -167,6 +167,41 @@ public class CLDRFile implements Freezable<CLDRFile>, Iterable<String>, LocaleSt
 
     private File supplementalDirectory;
 
+    /**
+     * Does the value in question either match or inherent the current value?
+     *
+     * <p>To match, the value in question and the current value must be non-null and equal.
+     *
+     * <p>To inherit the current value, the value in question must be INHERITANCE_MARKER and the
+     * current value must equal the bailey value.
+     *
+     * <p>This CLDRFile is only used here for getBaileyValue, not to get curValue
+     *
+     * @param value the value in question
+     * @param curValue the current value, that is, XMLSource.getValueAtDPath(xpathString)
+     * @param xpathString the path identifier
+     * @return true if it matches or inherits, else false
+     */
+    public boolean equalsOrInheritsCurrentValue(String value, String curValue, String xpathString) {
+        if (value == null || curValue == null) {
+            return false;
+        }
+        if (value.equals(curValue)) {
+            return true;
+        }
+        if (value.equals(CldrUtility.INHERITANCE_MARKER)) {
+            String baileyValue = getBaileyValue(xpathString, null, null);
+            if (baileyValue == null) {
+                /* This may happen for Invalid XPath; InvalidXPathException may be thrown. */
+                return false;
+            }
+            if (curValue.equals(baileyValue)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
     public enum DraftStatus {
         unconfirmed,
         provisional,
